requirements.txt
python-telegram-bot==20.7
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.121.0
langchain==0.1.12
langchain-google-genai==0.0.11
python-dotenv==1.0.1
jdatetime==4.1.1
pytz==2024.1
google-generativeai==0.3.2 